Fixed up the Index table
This is starting to look good, here's a screenshot http://imgur.com/rMUzD

Basically had to play with the CSS for an hour in order to really
understand what do do. Lucking SO came to the rescue as it usually does.
Next job is to play with the database's CDATA and figure out what's making
the newlines.