Added a Year action where it filters movies by year
It does that by making an SQL query with the db.Movies.SqlQuery() method.
I'm not exactly sure if that's correctly done, but it's working for now.
Maybe I'll add a todo.

Another thing I did was fix the POX I got from the Netflix API, because
normally it'd crash once it hit a line in the POX that wasn't a proper xml
line, so I just went through the POX file and Joined all the lines that
weren't properly formatted, due to the way CDATA was returned, and then
added it all to the DB. So the db went from under a MB to 9MB, or so. 7k
items to 55k. It takes about 25 minutes to go through the file and add
them all, so I think we'll have to optimize the data we're storing, since
a lot of it is repetitive, but we can leave that for later.

If you get a funny looking table on the Year page, it's because the genres
are too tall, so just make the table width wider, and you'll be good. CSS
is going to be a pain. But until it gets too ridiculous I think we'll be
in good shape, as this hasn't been to hard yet.

Except for the SQL business. F that noise.