Added some weak dropdown lists, and year filters if I haven't already
Fucking hell dude, that's hard to read, the docs I mean. You need to
create a List<SelectListItem>, then fill it with Items, then pass it to a
SelectList constructor AND THEN tell it which fields are values and which
are texts. How is that not done automatically. Holy shit, the docs are
garbage. Too friggin long I was working on this when it should be plainly
obvious.

Whatever, learning experience.

Also played with GETs and POSTs for pages, but since Chrome caches stuff,
I think I'll need to figure out a way to tell the browser that I've
updated the page, since either a GET or a POST call would result in a
different View. Of course I've scrapped that for now.

2 hours figuring all that out! Otherwise though MVC4 has been fun.