Holy tits, huge change to database
split up the Movie into smaller bits: MovieToGenre and Boxart and then
added a Genre model too. Movie point to the BoxArtId that holds all the
cols of img href string. need to figure out the genres thing as I will
need a query to pull the associated genres to the movie by using the
movieid as a key to grab all the linked genreIDs.

made slight changes to the API controller where it also looks for genres,
boxarts and movieToGenres