It runs but not properly, missing genre and boxart data
It's because I was writing code in the API action, instead of the Full
action. This will teach me to try to use logic written in the controller
and avoid using functions. Will need to try to use more functions and
properly organize my funcs from now on.

Will try to move it around after this.