Database is almost done now, boxarts and MtG works
Removed the boxart col from Movies since you can just search the movie id
in the Boxarts table and get it there. I don't think there'd be a
difference in speed between looking up a PK and a regular column, but I
could be wrong. About to build the db

TODO: improve tracing, so I know how far along the table is moving when
I'm building
	and set up exception handling. It crashes when there's an empty
	genre table, simple check to make sure there's populated genre
	table I think. Could even make it build it before it tries to use
	it eh? That'd be clever.