I DID IT! Database is properly populated!
Only takes about 50 minutes, but it should be quicker to poll and stuff
right, since it's all split up? I dunno, about to test it all, but I
really hope it does. Preliminary look says it's good at the raw tables
looks good. The trick was to disable DbContext.Configuration.DetectChanges
or something similar. Pretty fun stuff.

Added a helper function to build a list of checkpoints, so like you give
it say 10 and say you want 5 checkpoints, it'll spit out a list full of
ints 0,2,4,6,8,10 to make up 5 check points. It's a bit funky in that
there's no error checks or anything but it's working for now.