Fixed the Movies/Index Count property
It's because I didn't understand the difference about DbContext.Database
and DbContext.Tables, see my answer on SO

understand that dbContext.Database.SqlQuery<T>() isn't the same as
dbContext.TableName.SqlQuery(). The difference is that you're calling it
on Database and feeding it a type, rather than a table and no type.
Confused me a fair bit.

http://stackoverflow.com/questions/13031281/filling-child-entity-with-entity-framework-sqlquery/13039222#comment18965867_13039222