Added a Parameter Reflection
Added the ability to write out the possible parameters that you can pass
to the Controller. Makes it easier for us to use I think. It doesn't do
much other than list the name though. If we need to make it more complex
than that, I think we'll have to work on it more.