DID IT! Made a new view model for matching movies and its genres
Now you can go to Movies/Genres and look up a genre and you'll find all
movies that match it.

The way I did it was a 2 levelled SQL query to grab all the MovieToGenre
FKs along with the string for the genre that matched a given searched
genre, and then put all the genre strings into a dict's key, named
movie_id.

So then you go dict[movie_id] and you'll have all the genres in a nice
list. From there, it was just a matter of creating a ViewModel from the
list and Movie.

It's actually pretty ugly, to have to go model.movie.movie_id, instead of
just movie.movie_id, but it'll do for now.

Whatever the case, I'm pretty stoked. Will have to figure out how to make
the list GETted nicely, because right now the View just has them all space
joined, rather than comma joined.

Hopefully I'll improve on it tonight, but I don't think that'll happen.