Added headers to the filter printout
I think I'mna need to add a log to the site, so that I can monitor what's
going on. Right now I'm blind as a bat, and I think that should change
fairly quickly.

I'll need to decide on how I should track users across pages and things
like that.