Almost done marrying the omdbentries to the FilterMovie method
Will have to either create a new View that'll accept NITVMs or tweak
Genres.cshtml and break everything.... It's tough, so I thikn I'll just
make a new Genre copy and a new method similar to test i guess. It'll suck
to do again, since that's what /test was, but hey that's development for
you. Either way though, it's good that progress is moving  along. Having a
deadline is really good.