Moved the /sort logic mostly to Tools
Created a method for pulling all the movies from the DB so you can just
rely on that instead of having to rewrite it all the time.

What's left is the OrderBy stuff.

NOTE: if you don't pass the dbContext to the method I made, it'll take
longer to run, not sure why.