Preliminary Tag Work Complete. I have a model and view model that works lovely thanks to tankorsmash. movie will show all tags it is tagged with and on hover it will show how many times a particular tag has been used/
Need to add: Click on tag to goto tag page which will display the users
who have used this tag, and what movies are tagged with it. You shoudl
then be able to goto user page or movie page from there.

this means i need to add a tag controller that takes a parameter of String
tagName