Added a comment to the bottom of Genres.cshtml
Need clean the shit out of /testsort though, especially after Tilsons code
was pretty much spot less. Oh well, put in a good day today. I'll probably
get back to it tonight. I've got a bit of a game plan for the pagination
stuff. If I get the time, I can also add in the year and rating range,
because that's an essential part of the site.

I think we'll have to either delay the site publishing, or make it clear
that this is an open beta for the site, since a lot of the functionality
isn't quite polished and I'm sure we will be able to improve the page
loading and stuff, along with taking user feedback into consideration.

We'll probably have to create a feedback@watw.com to handle that stuff.