Added page numbers. Realized the downside to the List sorting, over Linq:
As you increment pages, you have to pull more and more items off the list
all at once. Say if it's page 150 (which is possible, as the max page
count is 551) is that you pull 150x28 movies only to discard 95% of them.
Basically we'll have to go back to the old OrderBy way and just hope that
the server is faster than Tilson's computer, and at least as fast as mine.
I'm sure it is...right?