Tried and failed some fancy shadows on Details2.cshtml
It's wasn't working as painlessly as I was hoping (that is to say entirely
painless) so I'm stopping work for the night. It's 11:20, that's about 4
hours now, and 4 hours today so fuck it, thats a full work day if you ask
me!

Made some good progress, idgaf.

So I'm probably going to have to go back to OrderBy from LINQ __or__ have
it so /Movies/Index doesn't show all the movies ever, but instead shows
like the first highest rated one, or something. No sense having them see
the literally slowest page in the site on their first visit. Maybe I'll
set up a welcome page with a nice example of how to use the site.