Solved strange pagination issue
in the end it was because I was using movies currently on the page as if
it was a constant, so I just needed to make sure it didn't change. Jolly
good show. Will need to work on the divide by zero error next, which is
just a matter of making sure that the page number isn't out of range
before I return the result