Separated the ViewBag setting for pagination into its own method
It's not a very good one mind you, but it's there. Want to change the
pagination so it doesn't print out 2k links, but the process seems a bit
tricky.

I think what I'll do is to try to make it so there's a maximum of eleven
links shown, not including first, prev, next, and last. For that I'll need
to test for total pages, and show as close to 5 before as I can, then fill
the rest of the pages after the current one.