Added a little comment to explain the strange looking if-stmt
In the pagination, that is.

The real reason I commited this is because I want to write down the
thought process here for the new pagination strategy. I'm going to fill a
ViewBag full of page numbers, about 11, so in a normal case, it'll be the
5 before, 6 is current and the last set are the 5 following pages. I'll
have it so the prev, next buttons are auto generated based on the current
page, and then I'll have to find the first and last page of the total
pages.

Still can't picture whether page 0 is the start or not...