Fixed the 5/1/6 page quirk
The out of bounds error wasn't a problem because some code I had already
set in place takes care of that for me. Can't remember exactly what code
that is though.

Next up, change the Site.css to make the pagination look a little better,
since we're going to try to publish tomorrow.