OptimizedPopulateOmdbTable is working at 20%
The imdb stuff can be entered pretty quickly, need to match the RT  data
and then match the OE to the movies.

I really want to fix the pretty nasty intial db creation. As it stands,
you need to go to /movies or sometihng to intialize the membership
database, then go to /user/seed to create the admin account THEN you can
create the FullDbBuild thing. Pretty unhappy with that, let me say :(

I'll see what I can do in the next little bit, or tomorrow.