Added maturity rating to Index
Working on adding it to the results. Not sure how I want to do it though.
Could make it so it sends an integer back to the server, or sends a string
that I parse. If it's a string I parse, it'll be easier for the user to
custom send the param, but if it's only an int, it'll be easier to convert
the string param to an int without worrying if it's a correct int.

I think I'll just make it so there's a list of translation from MR to
MPAA, because it needs to get done anyway.