Fixed the RT data misalignment
By manually assigning the PK for a table. I figure it's an okay thing to
do because we already have checks in place to make sure we don't add the
same data twice anyway, and this'll allow us to base our OmdbEntry table
off the OMDBAPI which is surely better organized that our db anyways.

JT if you've got a solution for this though, I'm all ears