Oh the things I have changed
1. Netflix xhanged the urls for the boxarts crom cdn-0 to cdn0 and instead
of .com it is now .net. There was soem programming that was stripping the
url parts out of the url form netflix and putting it back in later. I
removed all of this as netflix alread gives us the url for the boxart and
I am using it as is. so far so good.

2. Changed avg_rating in the database to double instead of string. this
makes queries against that field easier to handle as there is less data
being TryParsed. This also works well so far with testing.

Still need to do:

1. add dynamic jquery validaiton for search box depending on what is
selected in the select drop down

2. FIX THE GODDAMN CSS.